feat: add genres to film dao

+++ src/main/java/ru/yandex/practicum/filmorate/dao/FilmDbStorage.java
@@ -5,26 +5,18 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DataRetrievalFailureException;
-import org.springframework.dao.NonTransientDataAccessResourceException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.jdbc.support.SQLErrorCodeSQLExceptionTranslator;
 import org.springframework.stereotype.Repository;
-import ru.yandex.practicum.filmorate.exception.ResourceNotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.MpaRating;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.GenreStorage;
-import ru.yandex.practicum.filmorate.storage.MpaRatingStorage;
 
-import javax.script.ScriptException;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Repository
 @Qualifier("FilmDbStorage")
@@ -33,8 +25,6 @@ import java.util.Optional;
 public class FilmDbStorage implements FilmStorage {
 
     private final JdbcTemplate jdbcTemplate;
-    private final MpaRatingStorage mpaRatingStorage;
-    private final GenreStorage genreStorage;
 
     @Override
     public Film create(Film film) {
@@ -44,7 +34,7 @@ public class FilmDbStorage implements FilmStorage {
                     .usingGeneratedKeyColumns("id");
             long filmGeneratedId = simpleJdbcInsert.executeAndReturnKey(film.toMap()).longValue();
             film.setId(filmGeneratedId);
-            putGenres(film);
+            saveFilmGenres(film);
             log.info("Фильм {} c id={} успешно добавлен", film.getName(), film.getId());
         } catch (DataAccessException e) {
             log.warn("Ошибка получения жанра или MPA");
@@ -58,10 +48,15 @@ public class FilmDbStorage implements FilmStorage {
     public Optional<Film> get(Long id) {
         String sqlQuery =
                 "SELECT f.*, " +
-                        "mpa.id, mpa.name AS mpa_name " +
+                        "mpa.id, mpa.name AS mpa_name, " +
+                        "STRING_AGG(g.id, ', ') AS genre_id, " +
+                        "STRING_AGG(g.name, ', ') AS genre_name " +
                         "FROM films AS f " +
-                        "JOIN mpa_ratings AS mpa ON f.mpa_rating = mpa.id " +
-                        "WHERE f.id = ?";
+                        "LEFT JOIN mpa_ratings AS mpa ON f.mpa_rating = mpa.id " +
+                        "LEFT JOIN film_genre AS fg ON f.id = fg.film_id " +
+                        "LEFT JOIN genres AS g ON fg.genre_id = g.id " +
+                        "WHERE f.id = ? " +
+                        "GROUP BY f.id, f.name, f.description, f.release_date, f.duration, f.mpa_rating, mpa.name ";
         return jdbcTemplate.query(sqlQuery, this::mapRowToFilm, id).stream().findFirst();
     }
 
@@ -76,7 +71,7 @@ public class FilmDbStorage implements FilmStorage {
                 film.getDuration(),
                 film.getMpa().getId(),
                 film.getId());
-        putGenres(film);
+        saveFilmGenres(film);
 
         return film;
     }
@@ -91,20 +86,26 @@ public class FilmDbStorage implements FilmStorage {
     public List<Film> getAll() {
         String sqlQuery =
                 "SELECT f.*, " +
-                        "mpa.id, mpa.name AS mpa_name " +
+                        "mpa.id, mpa.name AS mpa_name, " +
+                        "STRING_AGG(g.id, ', ') AS genre_id, " +
+                        "STRING_AGG(g.name, ', ') AS genre_name " +
                         "FROM films AS f " +
-                        "JOIN mpa_ratings AS mpa ON f.mpa_rating = mpa.id ";
+                        "LEFT JOIN mpa_ratings AS mpa ON f.mpa_rating = mpa.id " +
+                        "LEFT JOIN film_genre AS fg ON f.id = fg.film_id " +
+                        "LEFT JOIN genres AS g ON fg.genre_id = g.id " +
+                        "GROUP BY f.id, f.name, f.description, f.release_date, f.duration, f.mpa_rating, mpa.name " +
+                        "ORDER BY f.id ";
 
         return jdbcTemplate.query(sqlQuery, this::mapRowToFilm);
     }
 
     @Override
-    public void addLike(Film film, User user) {
+    public void addLike(Long filmId, Long userId) {
 
     }
 
     @Override
-    public void deleteLike(Film film, User user) {
+    public void deleteLike(Long filmId, Long userId) {
 
     }
 
@@ -114,18 +115,42 @@ public class FilmDbStorage implements FilmStorage {
     }
 
 
-    private void putGenres(Film film) {
-        jdbcTemplate.update("DELETE FROM film_genre WHERE film_id = ?", film.getId());
-        jdbcTemplate.batchUpdate("INSERT INTO film_genre (film_id, genre_id) VALUES (?, ?)",
-                film.getGenres(),
-                film.getGenres().size(),
-                ((ps, genre) -> {
-                    ps.setLong(1, film.getId());
-                    ps.setLong(2, genre.getId());
-                }));
+    private void saveFilmGenres(Film film) {
+        if (film.getGenres() != null ) {
+            jdbcTemplate.update("DELETE FROM film_genre WHERE film_id = ?", film.getId());
+            jdbcTemplate.batchUpdate("INSERT INTO film_genre (film_id, genre_id) VALUES (?, ?)",
+                    film.getGenres(),
+                    film.getGenres().size(),
+                    ((ps, genre) -> {
+                        ps.setLong(1, film.getId());
+                        ps.setLong(2, genre.getId());
+                    }));
+        } else {
+            film.setGenres(new LinkedHashSet<>());
+        }
     }
 
     private Film mapRowToFilm(ResultSet resultSet, int i) throws SQLException {
+        LinkedHashSet<Genre> genres = new LinkedHashSet<>();
+
+        String genreIdString = resultSet.getString("genre_id");
+        String genreNameString = resultSet.getString("genre_name");
+
+        if (genreIdString != null && genreNameString != null &&
+                !genreIdString.isEmpty() && !genreNameString.isEmpty()) {
+
+            String[] genreIds = genreIdString.split(", ");
+            String[] genreNames = genreNameString.split(", ");
+
+            for (int j = 0; j < genreIds.length; j++) {
+                if (j < genreNames.length) {
+                    int genreId = Integer.parseInt(genreIds[j]);
+                    String genreName = genreNames[j];
+                    genres.add(new Genre(genreId, genreName));
+                }
+            }
+        }
+
         return Film.builder()
                 .id(resultSet.getLong("id"))
                 .name(resultSet.getString("name"))
@@ -133,6 +158,8 @@ public class FilmDbStorage implements FilmStorage {
                 .releaseDate(resultSet.getDate("release_date").toLocalDate())
                 .duration(resultSet.getInt("duration"))
                 .mpa(new MpaRating(resultSet.getInt("mpa_rating"), resultSet.getString("mpa_name")))
+                .genres(genres)
                 .build();
+
     }
 }
